Exception handling

+++ src/main/java/com/lucassa/ecomerce/services/ProductService.java
@@ -3,15 +3,16 @@ package com.lucassa.ecomerce.services;
 import com.lucassa.ecomerce.dto.ProductDto;
 import com.lucassa.ecomerce.entities.Product;
 import com.lucassa.ecomerce.repositories.ProductRepository;
+import com.lucassa.ecomerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -23,8 +24,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDto findBiId(Long id){
-        Optional<Product> product = productRepository.findById(id);
-        return  new ProductDto(product.get());
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
+        return  new ProductDto(product);
         //return modelMapper.map(product,ProductDto.class);
     }
 
@@ -48,19 +50,33 @@ public class ProductService {
 
     @Transactional
     public ProductDto update(Long id, ProductDto dto){
-        Product product = productRepository.getReferenceById(id);
-        product.setName(dto.getName());
-        product.setDescription(dto.getDescription());
-        product.setPrice(dto.getPrice());
-        product.setImgUrl(dto.getImgUrl());
-        product = productRepository.save(product);
+       try {
+           Product product = productRepository.getReferenceById(id);
+           product.setName(dto.getName());
+           product.setDescription(dto.getDescription());
+           product.setPrice(dto.getPrice());
+           product.setImgUrl(dto.getImgUrl());
+           product = productRepository.save(product);
 
-        return new ProductDto(product);
+           return new ProductDto(product);
+
+       }catch (EntityNotFoundException e){
+          throw new ResourceNotFoundException("Recuso não encontrado");
+       }
+
     }
 
-    @Transactional
-    public void delete(Long id){
-        productRepository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
